Extract chat empty state renderer
