+++ service-worker.js
@@ -78,6 +78,7 @@ const APP_SHELL = [
   '/js/chat-panel.js',
   '/js/chat-discussion.js',
   '/js/chat-onboarding.js',
+  '/js/chat-empty-state.js',
   '/js/chat-render.js',
   '/js/chat-send.js',
   '/js/chat-icons.js',
